Update Exercise and Goal models and controllers
This commit updates the Exercise and Goal models and controllers in the server. The Exercise model now exports the ExerciseSchema instead of the Exercise model itself, and the Goal model exports the GoalSchema instead of the Goal model itself. This change improves the modularity and reusability of the codebase. Additionally, the Goal import statement in the server index.js file has been updated to use the file extension. These updates enhance the maintainability and organization of the application.