Add Goals page
This commit adds a new page called Goals to the application. The Goals page displays a list of goals fetched from the server using the fetchGoals function. Each goal is rendered as a list item with its title and description. This feature allows users to view and manage their goals within the application.